Split GCE/GKE upgrade mechanics to a separate file
